khuyenmai

+++ Web_Client/Services/ISanphamService.cs
@@ -4,7 +4,7 @@ namespace Web_Client.Services
 {
     public interface ISanphamService
     {
-        string FormatVND(int price);
+        string FormatVND(double price);
         Task<List<Sanpham_Model>> GetAllProduct();
         string GetDetail(string id);
         Task<Sanpham_Model> GetProduct(string id);
